test_py/requirements.txt
#sudo apt install python3-dev libpq-dev gcc libc-dev libffi-dev
#apk add python3-dev libpq-dev gcc libc-dev libffi-dev
pandas
numpy
requests
psycopg2
sqlalchemy
boto3
jwt
flask
flask-kafka
flasgger
kafka-python
gunicorn
